core/framebuffer: use i32 rather than usize
Use i32 rather than usize, in order to match the types in the C++ template.

+++ src/graphics/linedrawer.rs
@@ -15,7 +15,7 @@ fn draw_x_line(
     let slope = dir as f32 * (ey as f32 - sy as f32) / (ex as f32 - sx as f32);
 
     while x != ex {
-        fb.plot_pixel(x as usize, y as usize, 1.0, 1.0, 1.0)?;
+        fb.plot_pixel(x, y as i32, 1.0, 1.0, 1.0)?;
         y += slope;
         x += dir;
     }
@@ -38,7 +38,7 @@ fn draw_y_line(
     let slope = dir as f32 * (ex as f32 - sx as f32) / (ey as f32 - sy as f32);
 
     while y != ey {
-        fb.plot_pixel(x as usize, y as usize, 1.0, 1.0, 1.0)?;
+        fb.plot_pixel(x as i32, y, 1.0, 1.0, 1.0)?;
         x += slope;
         y += dir;
     }
@@ -55,7 +55,7 @@ pub fn draw_line(
     ey: i32, // End Y coordinate.
 ) -> Result<(), &'static str> {
     if sx == ex && sy == ey {
-        fb.plot_pixel(sx as usize, sy as usize, 1.0, 1.0, 1.0)
+        fb.plot_pixel(sx, sy, 1.0, 1.0, 1.0)
     } else if (ex - sx).pow(2) >= (ey - sy).pow(2) {
         draw_x_line(fb, sx, sy, ex, ey)
     } else {
